Fix issue with setting "scale" attribute in IE (#287).
In IE setting the scale attribute with cx or cy set to 0 would
not work properly because the 0 would be converted to null before
the actual scale function was called. The fix is to use the same
code as is used for SVG browsers for converting the attr string
value to function arguments.