Added caching of fill patterns by url.
I noticed in Firefox that dyamically changing the pattern fill for a path
(e.g. fill: "url(image.jpg)") seemed slow, and caused the image to flash as
its fill was removed and re-added.  Investigating I noticed the pattern was
re-created with a new UUID each time the fill was changed, even when the URL
was the same. Caching the patterns speeds things up greatly.