isPointInside() used to use the cached "realPath" var to calculate a return. This meant if you modified a shapes attributes using the .attr() method (as opposed to using a transform) this meant the real path used would be wrong.
This forces isPointInside to always re-calculate the realPath.

Looking through the code, it actually looks like realPath is only used as a local variable in a few methods - and not as a cached "global" variable. It might be worth removing it entirely - or alternatively create a full feature global to cache the realPath.