Use 'none' instead of '0' for 'stroke-dasharray'.
Certain browsers have begun interpreting SVG 'stroke-dasharray' with a
value of '0' as *do not render a stroke*. The recommended value is to
use 'none'.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1149516 for Firefox
37.

This also causes invalid 'stroke-dasharray' values to be interpreted as
'none' (which is inline with how the VML is interpreted).