Fix drag offsets passed to "drag.start" from touch drags in elproto.drag
In `elproto.drag`, `x` and `y` are initialized to `e.clientX` and `e.clientY` respectively, but then there is code to update those values for touch events. Ultimately, the updated `x` and `y` values are added to `scrollX/scrollY` and then stored in `this._drag.x/this._drag.y`. The values passed along to `eve("raphael.drag.start.", ...)` were `e.clientX + scrollX/e.clientY + scrollY` which are the pre-corrected values. With this PR, the corrected values `this._drag.x/this._drag.y` are passed instead.